Improve Makefile.
This annoyed the hell out of me for a while, because my shoddy scripts
before would always force a recompilation of the target libraries
because it couldn't keep track of timestamps. This refactor makes it
such that if I type `make` twice, it doesn't recompile twice - just once.

It's not an end-all-be-all, granted, but it'll do.