Add Vagrantfile.
I recently switched my computer to a mbp; that's one of the reasons why
there's been a noticeable lack of progress lately. I still have my
other laptop, but... its battery life is terrible and it is literally
falling to pieces. </lifestory>

Anyway, this is good also, because it provides a consistent dev
environment to use. So use it.